Clean up codebase: remove old files, add SFL heuristics and RL training pipeline
- Removed old/outdated files (OFC-Poker-Agents, old policy nets, old docs)
- Added clean SFL heuristics tools (simulate_sfl_stats, tune_sfl, etc.)
- Added supervised learning pipeline (train_supervised_sfl.py)
- Added RL training pipeline (train_rl_policy.py, evaluate_rl_policy.py)
- Added dataset generation script (generate_sfl_dataset.py) with parallel support
- Updated README with current architecture
- Fixed C++ step_state bugs and SFL scoring integration

+++ cpp/ofc_encode.cpp
@@ -17,7 +17,7 @@ namespace py = pybind11;
 // 682..837: draw 3 * 52 one-hots (contiguous blocks)
 
 py::array_t<float> encode_state_batch_ints(py::array_t<int16_t> boards,  // (N,13)
-                                           py::array_t<int8_t> rounds,   // (N,)
+                                           py::array_t<int16_t> rounds,   // (N,)
                                            py::array_t<int16_t> draws,   // (N,3)
                                            py::array_t<int16_t> deck_sizes) { // (N,)
   auto B = boards.unchecked<2>();
